+++ src/providers/dankerino/DankerinoBadges.cpp
@@ -34,8 +34,8 @@ boost::optional<EmotePtr> DankerinoBadges::getBadge(const UserId &id)
 
 void DankerinoBadges::loadDankerinoBadges()
 {
-    static QUrl url(
-        "https://raw.githubusercontent.com/Mm2PL/chatterino2/dankerino/badges.json");
+    static QUrl url("https://raw.githubusercontent.com/Mm2PL/chatterino2/"
+                    "dankerino/badges.json");
 
     NetworkRequest(url)
         .onSuccess([this](auto result) -> Outcome {