+++ src/providers/twitch/api/Helix.hpp
@@ -192,6 +192,11 @@ public:
                       std::function<void()> successCallback,
                       HelixFailureCallback failureCallback);
 
+    void updateChannel(QString broadcasterId, QString gameId, QString language,
+                       QString title,
+                       std::function<void(NetworkResult)> successCallback,
+                       HelixFailureCallback failureCallback);
+
     void update(QString clientId, QString oauthToken);
 
     static void initialize();